refactor(clickclack): split discussion service workflows
